feat: Transform NSColor.Tokens to ColorToken
